- removed the illogical requirement that overflow:hidden be used on scroll areas
This makes the plugin have fewer requirements and be much more of a drop in solution as you don't need to go around setting all your scroll areas to overflow hidden when you first add it to your app. This also makes more sense in light of the new option to disableIfOSBarsOverlay as in iOS or common configurations of Mac OS which are often preferable to JS sliders.