crossdomain issue fix
`window.getComputedStyle` doesn't work in Firefox when embedding perfectscrollbar in some page in an iframe on some other domain/subdomain. it cannot access window from there. it throws a "`Given URL is not allowed by the Application configuration`" error and expects some special headers.

when using `element.style` there are no errors and no need for extra cross-domain headers with the plugin.

please note that this happened only on firefox, it worked fine in chrome