Fix Scroll Propagation for Touchpads
Original implementation caused choppy behavior when hitting container
bounds as it worked with delta===0 comparisons. This is perfectly fine
for wheels, but moving a finger perfectly straight on a touchpad is
unlikely, thus only a small percentage of scroll events were actually
propagated.

Improved implementation assumes the intended scroll axis by comparing
the absolute delta values of X and Y and propagates the events when
hitting the bounds on that axis.